+++ app/src/main/java/org/bidon/demoapp/ui/BannerScreen.kt
@@ -26,6 +26,7 @@ import androidx.navigation.NavHostController
 import kotlinx.coroutines.launch
 import org.bidon.demoapp.component.*
 import org.bidon.demoapp.ui.ext.toJson
+import org.bidon.demoapp.ui.ext.getImpressionInfo
 import org.bidon.sdk.ads.Ad
 import org.bidon.sdk.ads.AuctionInfo
 import org.bidon.sdk.ads.banner.BannerFormat
@@ -112,17 +113,16 @@ fun BannerScreen(navController: NavHostController) {
                                         if (showOnLoad.value) {
                                             bannerView?.showAd()
                                         }
+                                        logFlow.log("onAdLoaded ImpressionInfo: \n${ad.getImpressionInfo()}")
                                     }
 
-                                    override fun onAdLoadFailed(
-                                        auctionInfo: AuctionInfo?,
-                                        cause: BidonError
-                                    ) {
+                                    override fun onAdLoadFailed(auctionInfo: AuctionInfo?, cause: BidonError) {
                                         logFlow.log("onAdLoadFailed: $cause. AuctionInfo: \n${auctionInfo?.toJson()}")
                                     }
 
                                     override fun onAdShown(ad: Ad) {
                                         logFlow.log("onAdShown: $ad")
+                                        logFlow.log("onAdShown ImpressionInfo: \n${ad.getImpressionInfo()}")
                                     }
 
                                     override fun onAdClicked(ad: Ad) {
@@ -135,6 +135,7 @@ fun BannerScreen(navController: NavHostController) {
 
                                     override fun onRevenuePaid(ad: Ad, adValue: AdValue) {
                                         logFlow.log("onRevenuePaid: ad=$ad, adValue=$adValue")
+                                        logFlow.log("onRevenuePaid ImpressionInfo: \n${ad.getImpressionInfo()}")
                                     }
 
                                     override fun onAdShowFailed(cause: BidonError) {
@@ -232,10 +233,7 @@ fun BannerScreen(navController: NavHostController) {
                 Column(
                     modifier = Modifier
                         .padding(bottom = 2.dp)
-                        .background(
-                            MaterialTheme.colorScheme.secondary,
-                            MaterialTheme.shapes.medium
-                        )
+                        .background(MaterialTheme.colorScheme.secondary, MaterialTheme.shapes.medium)
                         .padding(4.dp)
                 ) {
                     Body2Text(text = logLine)